+++ day01/ex05/Human.cpp
@@ -30,7 +30,7 @@ std::string	Human::identify() const {
 
 	ss << this;
 	str = ss.str();
-	for (int i = 2; i < str.length(); ++i) {
+	for (unsigned long i = 2; i < str.length(); ++i) {
 		str[i] = toupper(str[i]);
 	}
 	return (str);
